refactor(inference): move the GP scoping rule to a leaf module
engine.py imports every agent at module level, and radio_agent instantiates its three transformer models AT import. Anything that wanted only the scoping rule paid for all of it: measured 16.7 s, with RoBERTa, the intent head, the NER model, the RAG agent and the orchestrator loaded into RAM and VRAM. src/agents/__init__ was made lazy precisely so touching one agent does not pull the family, and importing this helper from the hub undid that one layer up.

Its real dependencies are pandas, a logger and the GP-name resolver. From the leaf module the same import costs 0.39 s and loads no agent at all. engine re-exports it under the old private name, so the callers and tests that already reference engine._scope_laps_to_gp do not move.

A lap_state carrying `session_meta: null` also stopped raising: a present key holding None never triggers the two-arg get default, so the chained lookup raised where the honest answer is the loud unscoped fallback an unknown GP already takes.
